fix: change PORT placeholder

+++ configs/configs.go
@@ -19,7 +19,7 @@ type mongodb struct {
 	User     string
 	Password string
 	Host     string
-	Port     int16
+	Port     int
 	Database string
 }
